pyproject.toml
[project]
name = "python-sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "pyserial>=3.5",
    "python-dotenv>=1.1.1",
]
